# docs/api/openapi.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/api/v1/returns": {
            "post": {
                "operationId": "ReturnsController_create",
                "summary": "Create a sales return (original bill verified, open shift required, idempotent)",
                "parameters": [
                    {
                        "name": "idempotency-key",
                        "required": true,
                        "in": "header",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "Idempotency-Key",
                        "in": "header",
                        "description": "uuid v7 — mandatory; replays return the same document (no dup)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateReturnDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "returns"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "get": {
                "operationId": "ReturnsController_list",
                "summary": "List returns (newest first, cursor paginated)",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Envelope { data, meta }"
                    }
                },
                "tags": [
                    "returns"
                ]
            }
        },
...
        "/api/v1/reports/daily": {
            "get": {
                "operationId": "ReportsController_daily",
                "summary": "Daily sales aggregation (count / amount / vat / disc)",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Envelope { data, meta }"
                    }
                },
                "tags": [
                    "reports"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/reports/monthly": {
            "get": {
                "operationId": "ReportsController_monthly",
                "summary": "Monthly sales aggregation (count / amount / vat / disc)",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Envelope { data, meta }"
                    }
                },
                "tags": [
                    "reports"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/reports/by-item": {
            "get": {
                "operationId": "ReportsController_byItem",
                "summary": "Best-selling items (Arabic name, qty desc)",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Envelope { data, meta }"
                    }
                },
                "tags": [
                    "reports"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/reports/by-machine": {
            "get": {
                "operationId": "ReportsController_byMachine",
                "summary": "Sales per cashier machine (MACHINE_NO)",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Envelope { data, meta }"
                    }
                },
                "tags": [
                    "reports"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/reports/tax": {
            "get": {
                "operationId": "ReportsController_tax",
                "summary": "VAT/tax report by day (taxable base + VAT collected)",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Envelope { data, meta }"
                    }
                },
                "tags": [
                    "reports"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/reports/hourly-sales": {
            "get": {
                "operationId": "ReportsController_hourlySales",
                "summary": "Hourly sales distribution (00..23) from BILL_TIME",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Envelope { data, meta }"
                    }
                },
                "tags": [
                    "reports"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/reports/top-customers": {
            "get": {
                "operationId": "ReportsController_topCustomers",
                "summary": "Top customers by total sales (bill header key)",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Envelope { data, meta }"
                    }
                },
                "tags": [
                    "reports"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/reports/discount": {
            "get": {
                "operationId": "ReportsController_discount",
                "summary": "Discount report by day (header discount vs gross)",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Envelope { data, meta }"
                    }
                },
                "tags": [
                    "reports"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/reports/sales-by-category": {
            "get": {
                "operationId": "ReportsController_salesByCategory",
                "summary": "Sales grouped by item category (ITEM_TYPES)",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Envelope { data, meta }"
                    }
                },
                "tags": [
                    "reports"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/reports/z-report": {
            "get": {
                "operationId": "ReportsController_zReport",
                "summary": "Z-report: full end-of-shift/day close (single summary object)",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Envelope { data, meta }"
                    }
                },
                "tags": [
                    "reports"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/reports/slow-moving": {
            "get": {
                "operationId": "ReportsController_slowMoving",
                "summary": "Slow-moving items: least/zero-sold items in the period",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Envelope { data, meta }"
                    }
                },
                "tags": [
                    "reports"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/reports/profit": {
            "get": {
                "operationId": "ReportsController_profit",
                "summary": "Profit per item: revenue vs PRIMARY_COST (item master)",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Envelope { data, meta }"
                    }
                },
                "tags": [
                    "reports"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/reports/comparison": {
            "get": {
                "operationId": "ReportsController_comparison",
                "summary": "Two-period sales comparison (A vs B) with absolute + % deltas",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Envelope { data, meta }"
                    }
                },
                "tags": [
                    "reports"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/reports/item-movement": {
            "get": {
                "operationId": "ReportsController_itemMovement",
                "summary": "Full movement history of one item (sales + returns) in a period",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Envelope { data, meta }"
                    }
                },
                "tags": [
                    "reports"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/reports/audit": {
            "get": {
                "operationId": "ReportsController_audit",
                "summary": "Deleted-lines audit trail (IAS_POS_AUD_ITEM; POSR005 domain)",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Envelope { data, meta }"
                    }
                },
                "tags": [
                    "reports"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/reports/vat-detailed": {
            "get": {
                "operationId": "ReportsController_vatDetailed",
                "summary": "Detailed VAT report grouped by effective rate x item category",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Envelope { data, meta }"
                    }
                },
                "tags": [
                    "reports"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/reports/by-cashier": {
            "get": {
                "operationId": "ReportsController_byCashier",
                "summary": "Sales per cashier (MOTECH_POS bills+payments; POST012)",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Envelope { data, meta }"
                    }
                },
                "tags": [
                    "reports"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/reports/payment-methods": {
            "get": {
                "operationId": "ReportsController_paymentMethods",
                "summary": "Payment-method distribution (MOTECH_POS payments)",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Envelope { data, meta }"
                    }
                },
                "tags": [
                    "reports"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/reports/returns": {
            "get": {
                "operationId": "ReportsController_returns",
                "summary": "Returns aggregation by day (MOTECH_POS returns)",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Envelope { data, meta }"
                    }
                },
                "tags": [
                    "reports"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "info": {
        "title": "Motech POS API",
        "description": "NestJS backend (Oracle-first, Clean/Hexagonal)",
        "version": "0.2.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "bearer": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http"
            }
        },
        "schemas": {
...
            "CreateReturnLineDto": {
                "type": "object",
                "properties": {
                    "itemCode": {
                        "type": "string",
                        "example": "1060080003",
                        "description": "Item code (YSPOS23 I_CODE)"
                    },
                    "qty": {
                        "type": "number",
                        "example": 1,
                        "description": "Quantity to return (<= sold remaining)"
                    },
                    "unitPrice": {
                        "type": "number",
                        "example": 300,
                        "description": "Override price (else original)"
                    },
                    "vatPercent": {
                        "type": "number",
                        "example": 15,
                        "description": "Override VAT% (else original)"
                    },
                    "replaceAmount": {
                        "type": "number",
                        "example": 0,
                        "description": "Replacement amount (item swap)"
                    }
                },
                "required": [
                    "itemCode",
                    "qty"
                ]
            },
            "CreateReturnDto": {
                "type": "object",
                "properties": {
                    "originalBillNo": {
                        "type": "string",
                        "example": "26201300078",
                        "description": "Original bill being returned"
                    },
                    "cashierNo": {
                        "type": "number",
                        "example": 12,
                        "description": "Cashier number (must have open shift)"
                    },
                    "machineNo": {
                        "type": "number",
                        "example": 1
                    },
                    "returnType": {
                        "type": "number",
                        "example": 3,
                        "description": "RETURN_TYPE (default 3 = cash refund)"
                    },
                    "currency": {
                        "type": "string",
                        "example": "YER"
                    },
                    "clientOperationId": {
                        "type": "string",
                        "example": "018f2c1a-...uuidv7",
                        "description": "Client op id"
                    },
                    "lines": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CreateReturnLineDto"
                        }
                    }
                },
                "required": [
                    "originalBillNo",
                    "cashierNo",
                    "lines"
                ]
            },
...
        }
    }
}